refazendo o front v1 (navbar mobile responsiva)

+++ TCM/Repositorio/IProdutoRepositorio.cs
@@ -33,6 +33,6 @@ namespace TCM.Repositorio
         void RemoverComentario(int comentId);
         IEnumerable<Comentario> TodosComentarios();
         IEnumerable<Comentario> ComentariosProduto(int produtoId);
-        int TotalAvaliacoes(int produtoId);
+        double TotalAvaliacoes(int produtoId);
     }
 }
